requirements.txt
pillow==10.2.0
pytube==15.0.0
